fix: 🧩 修复横向布局下最大化失效
