+++ src/layouts/components/Main/components/Maximize.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script setup lang="ts">
-import { computed, watch } from "vue";
+import { computed } from "vue";
 import { GlobalStore } from "@/stores";
 
 const globalStore = GlobalStore();
@@ -13,17 +13,6 @@ const themeConfig = computed(() => globalStore.themeConfig);
 const exitMaximize = () => {
 	globalStore.setThemeConfig({ ...themeConfig.value, maximize: false });
 };
-
-// 监听当前页是否全屏，动态添加 class
-watch(
-	() => themeConfig.value.maximize,
-	() => {
-		const app = document.getElementById("app") as HTMLElement;
-		if (themeConfig.value.maximize) app.classList.add("main-maximize");
-		else app.classList.remove("main-maximize");
-	},
-	{ immediate: true }
-);
 </script>
 
 <style scoped lang="scss">
